pyproject.toml
[project]
name = "projeto-analise-sdr"
version = "2.0.0"
description = "SDR Analytics Dashboard - Sistema de Análise de Atendimento"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Gabriel Pastega"}
]
license = "MIT"
keywords = ["sdr", "analytics", "dashboard", "streamlit"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "pydantic>=2.10.5",
    "pytz>=2024.1",
    "httpx>=0.28.1",
    "streamlit>=1.41.0",
    "plotly>=6.1.1",
    "pandas>=2.2.3",
    "google-cloud-bigquery>=3.28.0",
    "python-dotenv>=1.0.1",
    "nest-asyncio>=1.6.0",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.45",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-pptx>=1.0.2",
    "psycopg2-binary>=2.9.11",
    "alembic>=1.17.2",
    "requests>=2.32.0",
    "google-genai (>=1.56.0,<2.0.0)",
    "streamlit-echarts (>=0.4.0,<0.5.0)",
    "streamlit-google-auth (>=1.1.8,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "sentry-sdk (>=2.48.0,<3.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.8",
    "mypy>=1.19.0",
    "pytest-cov>=7.0.0",
    "pre-commit>=4.5.0",
    "bandit>=1.9.2",
    "mkdocs-material>=9.7.0",
    "types-pytz>=2025.2.0.20251108",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E402"]  # Imports after load_dotenv() are intentional

[tool.mypy]
python_version = "3.14"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
filterwarnings = [
    "ignore::DeprecationWarning:google.genai.*",
]
